fix(gb): read_rom takes a ROM offset, so it cannot be 16-bit
read_rom's address parameter was uint16_t, but every caller passes a ROM
OFFSET rather than a Game Boy address:

    read_rom(..., (address - 0x4000) + (gb_cart->rom_bank * 0x4000), ...)

Bank 4 alone is 0x10000, so from there upward the offset wrapped. A 1 MB
cartridge read banks 0-3 correctly and returned bank & 3 for all the rest.

Nothing about that looks like a fault from outside. The header is in bank 0,
so it checksums; the pak handshake, the pak's own bank window, the MBC bank
selects and the whole 1 MB stream all complete and look right in a trace. The
game only finds out at the end, when the cartridge's global checksum does not
match, and Pokemon Stadium reports that as "The Transfer Pak is not set
properly. Please turn the N64 Control Deck OFF and check all connections" --
which reads as a seating fault rather than a truncated read.

Measured: with this widened to uint32_t, that error stops appearing at the
point the GB Tower loads a cartridge. Verified by photographing the tower,
since the failure is a message on screen that no log line reports.

The other uint16_t address parameters here are genuine Game Boy addresses and
stay as they are; read_ram's largest offset is 0x7fff and also fits. read_rom
is the only one handed a ROM offset.

+++ mupen64plus-core/src/device/gb/gb_cart.c
@@ -54,7 +54,14 @@ enum gbcart_extra_devices
 /* various helper functions for ram, rom, or MBC uses */
 
 
-static void read_rom(const void* rom_storage, const struct storage_backend_interface* irom_storage, uint16_t address, uint8_t* data, size_t size)
+/* `address` is a ROM OFFSET, not a Game Boy address: every caller passes
+ * (address - 0x4000) + rom_bank * 0x4000, which for bank 4 is already 0x10000.
+ * Declared uint16_t it wrapped, so a 1 MB cartridge read banks 0-3 correctly
+ * and returned bank & 3 for everything above -- the header checksummed, the
+ * whole ROM streamed, and the game rejected the cart at the end because its
+ * global checksum could not match. Pokemon Stadium reports that as "The
+ * Transfer Pak is not set properly". */
+static void read_rom(const void* rom_storage, const struct storage_backend_interface* irom_storage, uint32_t address, uint8_t* data, size_t size)
 {
     assert(size > 0);
